+++ src/main/java/cn/kizzzy/javafx/display/image/aoi/Vector3f.java
@@ -2,26 +2,26 @@ package cn.kizzzy.javafx.display.image.aoi;
 
 public class Vector3f extends Vector2f {
     
-    private double z;
+    private float z;
     
-    public Vector3f(double x) {
+    public Vector3f(float x) {
         this(x, 0);
     }
     
-    public Vector3f(double x, double y) {
+    public Vector3f(float x, float y) {
         this(x, y, 0);
     }
     
-    public Vector3f(double x, double y, double z) {
+    public Vector3f(float x, float y, float z) {
         super(x, y);
         this.z = z;
     }
     
-    public double getZ() {
+    public float getZ() {
         return z;
     }
     
-    public void setZ(double z) {
+    public void setZ(float z) {
         this.z = z;
     }
 }
