+++ src/main/java/cn/kizzzy/javafx/display/image/aoi/Vector4f.java
@@ -2,30 +2,30 @@ package cn.kizzzy.javafx.display.image.aoi;
 
 public class Vector4f extends Vector3f {
     
-    private double w;
+    private float w;
     
-    public Vector4f(double x) {
+    public Vector4f(float x) {
         this(x, 0);
     }
     
-    public Vector4f(double x, double y) {
+    public Vector4f(float x, float y) {
         this(x, y, 0);
     }
     
-    public Vector4f(double x, double y, double z) {
+    public Vector4f(float x, float y, float z) {
         this(x, y, z, 0);
     }
     
-    public Vector4f(double x, double y, double z, double w) {
+    public Vector4f(float x, float y, float z, float w) {
         super(x, y, z);
         this.w = w;
     }
     
-    public double getW() {
+    public float getW() {
         return w;
     }
     
-    public void setW(double w) {
+    public void setW(float w) {
         this.w = w;
     }
 }
